CHANGED: Offset type to long

+++ PacketTracerConfigUnlock/Targets.cs
@@ -12,7 +12,7 @@ namespace PacketTracerConfigUnlock
             (
                 "Cisco Packet Tracer 7.2.2 (x86)",
                 @"C:\Program Files (x86)\Cisco Packet Tracer 7.2.2\bin\PacketTracer7.exe",
-                new Dictionary<int, string>()
+                new Dictionary<long, string>()
                 {
                     {0x1063AA4, "E92B0100"},    // Unlock config tab
                     {0x16AAB0D, "E9530100"}     // Speed up loading times
@@ -22,7 +22,7 @@ namespace PacketTracerConfigUnlock
             (
                 "Cisco Packet Tracer 8.0.1 (x86)",
                 @"C:\Program Files (x86)\Cisco Packet Tracer 8.0.1\bin\PacketTracer.exe",
-                new Dictionary<int, string>()
+                new Dictionary<long, string>()
                 {
                     {0x112EC6D, "E9380200"},    // Unlock config tab
                     {0x191806D, "E9FF0600"}     // Speed up loading times
@@ -32,7 +32,7 @@ namespace PacketTracerConfigUnlock
             (
                 "Cisco Packet Tracer 7.2.2 (x64)",
                 @"C:\Program Files\Cisco Packet Tracer 7.2.2\bin\PacketTracer7.exe",
-                new Dictionary<int, string>()
+                new Dictionary<long, string>()
                 {
                     {0x147B638, "E9490100"},    // Unlock config tab
                     {0x1C58604, "E9650500"}     // Speed up loading times
@@ -42,7 +42,7 @@ namespace PacketTracerConfigUnlock
             (
                 "Cisco Packet Tracer 8.0.1 (x64)",
                 @"C:\Program Files\Cisco Packet Tracer 8.0.1\bin\PacketTracer.exe",
-                new Dictionary<int, string>()
+                new Dictionary<long, string>()
                 {
                     {0x1720076, "E9FA0200"},    // Unlock config tab
                     {0x218B564, "E9820800"}     // Speed up loading times
@@ -50,7 +50,7 @@ namespace PacketTracerConfigUnlock
             );
         }
 
-        private static void Patch(string name, string path, Dictionary<int, string> payloads)
+        private static void Patch(string name, string path, Dictionary<long, string> payloads)
         {
             if (!File.Exists(path)) { return; }
